+++ lib/baabedo/api_operations/search.rb
@@ -1,15 +1,17 @@
 module Baabedo
   module APIOperations
-    module List
+    module Search
       module ClassMethods
-        def all(params={}, opts={})
+        def search(query, params={}, opts={})
           opts = Util.normalize_opts(opts)
 
-          response, opts = request(:get, url, params, opts)
+          search_url = url(params) + '/_search'
+          params = params.merge({query: query})
+          response, opts = request(:get, search_url, params, opts)
           Util.convert_to_api_object(response, opts)
         end
 
-        def url
+        def url(params)
           return build_url(params) if respond_to?(:build_url)
           super
         end